# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bar Order Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type StaffRole = (typeof StaffRole)[keyof typeof StaffRole];

export const StaffRole = {
  admin: "admin",
  waitress: "waitress",
  bartender: "bartender",
} as const;

export interface StaffMember {
  id: number;
  name: string;
}

export interface AdminStaffMember {
  id: number;
  name: string;
  role: StaffRole;
}

export interface CreateStaffBody {
  name: string;
  role: StaffRole;
  /**
   * @minLength 4
   * @maxLength 4
   */
  pin: string;
}

export interface UpdateStaffBody {
  name?: string;
  role?: StaffRole;
  /**
   * @minLength 4
   * @maxLength 4
   */
  pin?: string;
}

export interface ChangePinBody {
  /**
   * @minLength 4
   * @maxLength 4
   */
  newPin: string;
}

export interface PinLoginBody {
  staffId: number;
  /**
   * @minLength 4
   * @maxLength 4
   */
  pin: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
  role?: StaffRole;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MenuItemBody {
  name: string;
  category: string;
  /** @minimum 0 */
  pricePence: number;
}

export interface MenuItem {
  id: number;
  name: string;
  category: string;
  pricePence: number;
}

export interface OrderItem {
  id: number;
  menuItemId: number;
  menuItemName: string;
  quantity: number;
  pricePence: number;
}

export type OrderBatchStatus =
  (typeof OrderBatchStatus)[keyof typeof OrderBatchStatus];

export const OrderBatchStatus = {
  pending: "pending",
  completed: "completed",
  paid: "paid",
  returned: "returned",
} as const;

export interface OrderBatch {
  id: number;
  customerName: string;
  waitressName: string;
  createdAt: string;
  completedAt?: string | null;
  status: OrderBatchStatus;
  correctionItemIds?: number[] | null;
  items: OrderItem[];
}

export interface CreateOrderItemBody {
  menuItemId: number;
  quantity: number;
}

export interface CreateOrderBatchBody {
  customerName: string;
  waitressName: string;
  items: CreateOrderItemBody[];
}

export interface Shift {
  id: number;
  staffId: number;
  staffName: string;
  startedAt: string;
  endedAt?: string | null;
}

export interface ShiftEnvelope {
  shift: Shift | null;
}

export interface ResubmitOrderBatchBody {
  items: CreateOrderItemBody[];
}

export interface EditOrderBatchBody {
  items: CreateOrderItemBody[];
}

export interface SettleWaiterBody {
  waitressName: string;
}

export interface SettleWaiterResult {
  waitressName: string;
  count: number;
  totalPence: number;
}

export interface ReturnOrderBatchBody {
  correctionItemIds: number[];
}

export type ActivityLogEntryDetails = { [key: string]: unknown } | null;

export interface ActivityLogEntry {
  id: number;
  timestamp: string;
  actorName: string;
  actorRole: string;
  action: string;
  details?: ActivityLogEntryDetails;
}

export type ActivityLogsResponse = ActivityLogEntry[];

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  /**
   * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
   */
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};

export type GetShiftsParams = {
  /**
   * Filter by date (YYYY-MM-DD). Defaults to today.
   */
  date?: string;
};

export type GetActivityLogsParams = {
  date?: string;
  actor?: string;
  action?: string;
};