# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Bar Order Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/staff:
    get:
      tags: [Admin]
      operationId: getAdminStaff
      summary: List all staff members with roles (admin only)
      responses:
        "200":
          description: List of all staff
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminStaffMember'
    post:
      tags: [Admin]
      operationId: createStaff
      summary: Create a new staff member (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStaffBody'
      responses:
        "201":
          description: Staff member created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminStaffMember'
  /admin/staff/{id}:
    patch:
      tags: [Admin]
      operationId: updateStaff
      summary: Update a staff member (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStaffBody'
      responses:
        "200":
          description: Staff member updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminStaffMember'
        "404":
          description: Not found
    delete:
      tags: [Admin]
      operationId: deleteStaff
      summary: Delete a staff member (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /menu-items:
    get:
      operationId: getMenuItems
      tags: [menu]
      summary: Get all menu items
      responses:
        "200":
          description: List of menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
    post:
      operationId: createMenuItem
      tags: [menu]
      summary: Create a new menu item (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemBody"
      responses:
        "201":
          description: Menu item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
  /menu-items/{id}:
    patch:
      operationId: updateMenuItem
      tags: [menu]
      summary: Update a menu item (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemBody"
      responses:
        "200":
          description: Menu item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
    delete:
      operationId: deleteMenuItem
      tags: [menu]
      summary: Delete a menu item (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted successfully
...
components:
...
  schemas:
...
    StaffRole:
      type: string
      enum: [admin, waitress, bartender]
...
    AdminStaffMember:
      type: object
      required: [id, name, role]
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          $ref: '#/components/schemas/StaffRole'
    CreateStaffBody:
      type: object
      required: [name, role, pin]
      properties:
        name:
          type: string
        role:
          $ref: '#/components/schemas/StaffRole'
        pin:
          type: string
          minLength: 4
          maxLength: 4
    UpdateStaffBody:
      type: object
      properties:
        name:
          type: string
        role:
          $ref: '#/components/schemas/StaffRole'
        pin:
          type: string
          minLength: 4
          maxLength: 4
...
    MenuItemBody:
      type: object
      required: [name, category, pricePence]
      properties:
        name:
          type: string
        category:
          type: string
        pricePence:
          type: integer
          minimum: 0
    MenuItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        pricePence:
          type: integer
      required:
        - id
        - name
        - category
        - pricePence
...